Rename list pagination to something more intuitive, and add full pagination to ADT table
